+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Added two new fired events, for suspending and resuming execution. (#637)
+- Added support for array arithmetic. (#628)
 
 ## [0.5.0] - 2021-04-15